move ticker to separate file, start refactoring
